refactor: Move data fetching logic to NewsFacade
