+++ src/components/main/news/NewsList.tsx
@@ -8,7 +8,7 @@ import { SwiperList } from '@/components';
 import { LinkWithArrow } from '@/components/ui/LinkWithArrow';
 import { Section } from '@/components/ui/Section';
 import { getLocale, getTranslations } from 'next-intl/server';
-import { fetchNews } from '@/services/newsFacade';
+import { fetchNews } from '@/facades/newsFacade';
 import { Tlocale } from '@/types';
 
 export const NewsList = async (): Promise<JSX.Element> => {
